Fix old, partially-grown plants.
They were appearing as "Unknown Item" because they were not registered
by the new farming mod code, so replace them with their new, fully-grown
versions.

While it would be nice to replace them with their equivalent
partially-grown versions, this is tricky because node data such as the
growing timers are not part of the old nodes; simply aliasing the old
version to the new thus results in the correct plant, but one that never
grows.  Aliasing old, partially-grown plants to the new, fully-grown
plant thus works around this issue and gives a small, one-time bonus to
anyone who has partially-grown plants.